# docs/swagger.yaml
basePath: /
definitions:
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.Policy:
    properties:
      action:
        type: string
      oldPolicy:
        items:
          type: string
        type: array
      policy:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a weave server api doc.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Weave Server API
  version: "2.0"
paths:
...
  /api/v1/namespaces/:namespace/:resource:
    get:
      consumes:
      - application/json
      description: List kube resource
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - JWT: []
      summary: List kube resource
      tags:
      - k8s
    post:
      consumes:
      - application/json
      description: Create kube resource
      parameters:
      - description: k8s json info
        in: body
        name: app
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - JWT: []
      summary: Create kube resource
      tags:
      - k8s
  /api/v1/namespaces/:namespace/:resource/:name:
    delete:
      consumes:
      - application/json
      description: Delete kube resource
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - JWT: []
      summary: Delete kube resource
      tags:
      - k8s
    get:
      consumes:
      - application/json
      description: Get kube resource
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - JWT: []
      summary: Get kube resource
      tags:
      - k8s
    put:
      consumes:
      - application/json
      description: Update kube resource
      parameters:
      - description: k8s json info
        in: body
        name: app
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - JWT: []
      summary: Update kube resource
      tags:
      - k8s
  /api/v1/policies:
    get:
      description: List rbac policy
      parameters:
      - description: 'ptype: p/g/g2'
        in: query
        name: ptype
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - JWT: []
      summary: List rbac policy
      tags:
      - rbac
    post:
      consumes:
      - application/json
      description: Handle rbac policy and storage
      parameters:
      - description: rbac policy info
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/model.Policy'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - JWT: []
      summary: Handle rbac policy
      tags:
      - rbac
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"